chore: save stuff in body commands
