+++ crates/types/src/motion_command.rs
@@ -219,6 +219,60 @@ pub enum ImageRegion {
     Top,
 }
 
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Deserialize,
+    Eq,
+    PartialEq,
+    Serialize,
+    PathSerialize,
+    PathDeserialize,
+    PathIntrospect,
+)]
+pub enum KickDirection {
+    Back,
+    Front,
+    Left,
+    Right,
+}
+
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Deserialize,
+    Eq,
+    PartialEq,
+    Serialize,
+    PathSerialize,
+    PathDeserialize,
+    PathIntrospect,
+)]
+pub enum KickVariant {
+    Forward,
+    Turn,
+    Side,
+}
+
+#[derive(
+    Clone,
+    Copy,
+    Debug,
+    Deserialize,
+    Eq,
+    PartialEq,
+    Serialize,
+    PathSerialize,
+    PathDeserialize,
+    PathIntrospect,
+)]
+pub enum JumpDirection {
+    Left,
+    Right,
+    Center,
+}
 
 #[derive(
     Clone,
