+++ crates/world_state/src/behavior/action.rs
@@ -28,6 +28,12 @@ pub fn prepare(blackboard: &mut Blackboard) -> Status {
     Status::Success
 }
 
+
+pub fn set_is_alternative(blackboard: &mut Blackboard) -> Status {
+    blackboard.is_alternative = true;
+    Status::Success
+}
+
 pub fn stand(blackboard: &mut Blackboard) -> Status {
     blackboard.body_motion = Some(BodyMotion::Stand);
     Status::Success
