+++ crates/world_state/src/behavior/mod.rs
@@ -2,8 +2,7 @@ pub mod action;
 pub mod behavior_tree;
 pub mod condition;
 pub mod head;
-pub mod kick_actions;
-pub mod kick_selector;
+pub mod kick;
 pub mod motion_assembler;
 pub mod node;
 pub mod switch_motion_type;
